requirements.txt
PyPDF2
Pillow
